+++ app/javascript/vue/tasks/asserted_distributions/new_asserted_distribution/components/Panel/PanelCitation.vue
@@ -21,8 +21,13 @@
         v-model:lock="store.lock.source"
         :klass="ASSERTED_DISTRIBUTION"
         :target="ASSERTED_DISTRIBUTION"
+        @update:modelValue="sendBroadcast"
         @lock="(e) => (store.lock.citation = e)"
-      />
+      >
+        <template #tabs-right>
+          <VBroadcast v-model="isBroadcastActive" />
+        </template>
+      </FormCitation>
     </template>
   </BlockLayout>
 </template>
@@ -31,11 +36,30 @@
 import { useStore } from '../../store/store'
 import { ASSERTED_DISTRIBUTION } from '@/constants'
 import { RouteNames } from '@/routes/routes.js'
+import { ref } from 'vue'
+import { useBroadcastChannel } from '@/composables'
 import BlockLayout from '@/components/layout/BlockLayout.vue'
 import FormCitation from '@/components/Form/FormCitation.vue'
 import VBtn from '@/components/ui/VBtn/index.vue'
+import VBroadcast from '@/components/ui/VBroadcast/VBroadcast.vue'
 
 const store = useStore()
+const isBroadcastActive = ref(false)
+
+const { post } = useBroadcastChannel({
+  name: 'citation',
+  onMessage({ data }) {
+    if (isBroadcastActive.value) {
+      store.citation = data
+    }
+  }
+})
+
+function sendBroadcast(data) {
+  if (isBroadcastActive.value) {
+    post(data)
+  }
+}
 
 function openTask() {
   window.open(RouteNames.NewSource)